thêm attribute product

+++ controllers/ProductController.js
@@ -1,5 +1,6 @@
 import { Sequelize, Op, where } from "sequelize";
 import db from "../models";
+import { getAvatarUrl } from "../helpers/imageHelper";
 
 
 export async function getProducts(req, res) {
@@ -17,9 +18,24 @@ export async function getProducts(req, res) {
     };
   }
 
+  // Khởi tạo whereClause cho thuộc tính động
+  let attributeWhereClause = {};
+  if (search.trim() !== "") {
+    attributeWhereClause = { value: { [Op.like]: `%${search}%` } }
+  }
+console.log(attributeWhereClause)
   const [products, totalProducts] = await Promise.all([
     db.Product.findAll({
       where: whereClause,
+      include: [
+        {
+          model: db.ProductAttributeValue,
+          as: "attributes",
+          include: [{ model: db.Attribute }],
+          where: attributeWhereClause,
+          required: false
+        },
+      ],
       limit: pageSize,
       offset: offset,
     }),
@@ -29,19 +45,30 @@ export async function getProducts(req, res) {
   ]);
   return res.status(200).json({
     message: "Lấy danh sách sản phẩm thành công",
-    data: products,
-    currentPage: parseInt(page, 10),
-    totalPages: Math.ceil(totalProducts / pageSize),
-    totalProducts,
+    data: products.map((product) => ({
+      ...product.get({ plain: true }),
+      image: getAvatarUrl(product.image),
+    })),
+    current_page: parseInt(page, 10),
+    total_pages: Math.ceil(totalProducts / pageSize),
+    total:totalProducts,
   });
 }
 
 export async function getProductById(req, res) {
   const { id } = req.params;
   const product = await db.Product.findByPk(id, {
-      include:[{ model: db.ProductImage, as: "product_images"}],
+    include: [
+      { model: db.ProductImage, as: "product_images" },
+      {
+        model: db.ProductAttributeValue,
+        as: "attributes",
+        include: [
+          {model: db.Attribute, attributes: ["name"]}
+        ]
+      },
+    ],
   });
-  console.log(product);
   if (!product) {
     return res.status(404).json({
       message: "Sản phẩm không tìm thấy",
@@ -49,38 +76,82 @@ export async function getProductById(req, res) {
   }
   res.status(200).json({
     message: "Lấy thông tin sản phẩm thành công",
-    data: product,
+    data: {
+      ...product.get({ plain: true }),
+      image: getAvatarUrl(product.image),
+    },
   });
 }
 
 export async function insertProducts(req, res) {
+  const { name, attributes = [], ...productData } = req.body;
   // Kiểm tra xem sản phẩm đã tồn tại chưa
-  const existingName = await db.Product.findOne({
-    where: { name: req.body.name },
+  const ProductExist = await db.Product.findOne({
+    where: { name },
   });
-  if (existingName) {
+  if (ProductExist) {
     return res.status(400).json({
       message: "Tên sản phẩm đã tồn tại, vui lòng chọn tên khác!",
     });
   }
   const product = await db.Product.create(req.body);
+
+  if (attributes.length > 0) {
+    for (const attr of attributes) {
+      const [attribute] = await db.Attribute.findOrCreate({
+        where: { name: attr.name },
+      });
+        await db.ProductAttributeValue.create({
+          product_id: product.id,
+          attribute_id: attribute.id,
+          value: attr.value,
+        });
+    }
+  }
   res.status(200).json({
     message: "Thêm mới sản phẩm thành công",
-    data: product,
+    data: {
+      ...product.get({ plain: true }),
+      image: getAvatarUrl(product.image),
+      attributes
+    },
   });
 }
 
 export async function deleteProducts(req, res) {
   const { id } = req.params;
+
+  // kiểm tra sản phẩm có trong orderDetail ko
+  const orderDetailExists = await db.OrderDetail.findOne({
+    where: { product_id },
+    include: [{
+      model: db.Order,
+      as: "order",
+      attributes: ["id", "status", "note", 'total', 'created_at']
+    }]
+  })
+
+  if (orderDetailExists) {
+    return res.status(400).json({
+      message: "Không thể xoá sản phẩm vì đã có đơn hàng tham chiếu đến sản phẩm này",
+      data: {
+        order: orderDetailExists.order
+      }
+    });
+  }
+
+  await db.ProductAttributeValue.destroy({
+    where: {id}
+  })
   const deleted = await db.Product.destroy({
     where: { id },
   });
   if (deleted) {
-    res.status(200).json({
+    return res.status(200).json({
       message: "xoá sản phẩm thành công",
     });
   } else {
-    res.status(200).json({
+    return res.status(400).json({
       message: "Sản phẩm không tìm thấy",
     });
   }
@@ -88,31 +159,43 @@ export async function deleteProducts(req, res) {
 
 export async function updateProducts(req, res) {
   const { id } = req.params;
-  const { name } = req.body;
-  if (name !== undefined) {
-    const existingProduct = await db.Product.findOne({
-      where: {
-        name,
-        id: { [Op.ne]: id },
-      },
-    });
-    if (existingProduct) {
-      return res.status(409).json({
-        message: "Tên sản phẩm đã tồn tại vui lòng chọn tên sản phẩm khác",
-      });
-    }
-  }
-  const updatedProduct = await db.Product.update(req.body, {
+  const { attributes = [], ...productData } = req.body;
+
+  const updatedRowCount = await db.Product.update(productData, {
     where: { id },
   });
-  if (updatedProduct) {
+
+  if (updatedRowCount) {
+    for (const attr of attributes) {
+      const [attribute] = await db.Attribute.findOrCreate({
+        where: { name: attr.name },
+      });
+      const productAttributeValue = await db.ProductAttributeValue.findOne({
+        where: {
+          product_id: id,
+          attribute_id: attribute.id,
+        },
+      });
+
+      if (productAttributeValue) {
+        await productAttributeValue.update({value: attr.value})
+      } else {
+        await db.ProductAttributeValue.create({
+          product_id: id,
+          attribute_id: attribute.id,
+          value: attr.value,
+        });
+      }
+    }
+
     return res.status(200).json({
-      message: "update sản phẩm thành công",
+      message: "Cập nhật sản phẩm thành công",
     });
-  }
-  {
-    return res.status(200).json({
+  } else {
+    return res.status(401).json({
       message: "Sản phẩm không tìm thấy",
     });
   }
+
+
 }
